requirements.txt
atproto>=0.0.48
pandas>=2.1.0
requests>=2.31.0
python-dateutil>=2.8.2